+++ addons/ezcha_network/dock/menu.gd
@@ -1,6 +1,6 @@
 extends ScrollContainer
 
 @onready var dock = get_parent()
-@onready var _ezcha: Node = get_node_or_null("/root/Ezcha")
+@onready var _ezcha: EzchaSingleton = get_node_or_null("/root/Ezcha")
 
 func _opened() -> void: pass