+++ addons/ezcha_network/lib/api/leaderboards.gd
@@ -15,49 +15,41 @@ var _update_mode_map: PackedStringArray = PackedStringArray(["set", "add", "subt
 
 ## Returns a paginated list of entries for a specific leaderboard.
 ## A session token is only required when attempting to access an unlisted leaderboard.
-func get_entries(leaderboard_id: String, page: int = 1, items_per_page: int = -1, session_token: String = "") -> EzchaPaginatedLeaderboardEntryListResponse:
-	var resp: EzchaPaginatedLeaderboardEntryListResponse = EzchaPaginatedLeaderboardEntryListResponse.new()
-	var builder: EzchaRequestBuilder = EzchaRequestBuilder.new()
-	builder\
+func get_entries(leaderboard_id: String, page: int = 1, items_per_page: int = -1, session_token: String = "") -> EzchaLeaderboardEntryListResponse:
+	return EzchaRequestBuilder.new()\
 		.set_method(HTTPClient.METHOD_GET)\
 		.set_endpoint("/v1/leaderboards/entries")\
 		.set_authentication(session_token)\
-		.set_response_object(resp)\
+		.set_response_object(EzchaLeaderboardEntryListResponse.new())\
 		.add_query_parameter("leaderboard_id", leaderboard_id)\
-		.add_query_parameter("page", page)
-	if (items_per_page > -1):
-		builder.add_query_parameter("items_per_page", items_per_page)
-	builder.fetch()
-	return resp
+		.add_query_parameter("page", page)\
+		.add_query_parameter("items_per_page", items_per_page if (items_per_page > -1) else null)\
+		.fetch()
 
 ## Updates a score from a game client using a session token.
 ## Requires a signing key to be configured.
 func post_entry_client(leaderboard_id: String, session_token: String, score: float, mode: UpdateMode = UpdateMode.SET) -> EzchaLeaderboardQueuedResponse:
-	var resp: EzchaLeaderboardQueuedResponse = EzchaLeaderboardQueuedResponse.new()
-	EzchaRequestBuilder.new()\
+	return EzchaRequestBuilder.new()\
 		.set_method(HTTPClient.METHOD_POST)\
 		.set_endpoint("/v1/leaderboards/entries/client")\
 		.set_authentication(session_token)\
 		.set_signing_key(_ezcha.get_signing_key())\
-		.set_response_object(resp)\
+		.set_response_object(EzchaLeaderboardQueuedResponse.new())\
 		.add_body_data("leaderboard_id", leaderboard_id)\
 		.add_body_data("score", score)\
 		.add_body_data("mode", _update_mode_map[mode])\
 		.fetch()
-	return resp
 
 ## Updates a score from a game server using an API key.
 ## Requires an API key to be configured.
 func post_entry_server(leaderboard_id: String, user_id: String, score: float, mode: UpdateMode = UpdateMode.SET) -> EzchaLeaderboardQueuedResponse:
-	var resp: EzchaLeaderboardQueuedResponse = EzchaLeaderboardQueuedResponse.new()
-	EzchaRequestBuilder.new()\
+	return EzchaRequestBuilder.new()\
 		.set_method(HTTPClient.METHOD_POST)\
 		.set_endpoint("/v1/leaderboards/entries/server")\
 		.set_authentication(_ezcha.get_api_key())\
-		.set_response_object(resp)\
+		.set_response_object(EzchaLeaderboardQueuedResponse.new())\
 		.add_body_data("leaderboard_id", leaderboard_id)\
 		.add_body_data("score", score)\
 		.add_body_data("mode", _update_mode_map[mode])\
 		.add_body_data("user_id", user_id)\
 		.fetch()
-	return resp