+++ addons/ezcha_network/lib/api/news.gd
@@ -5,15 +5,13 @@ class_name EzchaNewsAPI
 ## This should be accessed through the "Ezcha" singleton.
 
 ## Returns a paginated list of news posts based on the criteria provided.
-func get_list(page: int = 1, category: String = "", series: String = "", order: String = "") -> EzchaPaginatedNewsListResponse:
-	var resp: EzchaPaginatedNewsListResponse = EzchaPaginatedNewsListResponse.new()
-	EzchaRequestBuilder.new()\
+func get_list(page: int = 1, category: String = "", series: String = "", order: String = "") -> EzchaNewsListResponse:
+	return EzchaRequestBuilder.new()\
 		.set_method(HTTPClient.METHOD_GET)\
 		.set_endpoint("/v1/news/list")\
-		.set_response_object(resp)\
+		.set_response_object(EzchaNewsListResponse.new())\
 		.add_query_parameter("page", page)\
 		.add_query_parameter("category", category)\
 		.add_query_parameter("series", series)\
 		.add_query_parameter("order", order)\
 		.fetch()
-	return resp