+++ addons/ezcha_network/lib/responses/user_list.gd
@@ -1,11 +1,11 @@
-extends EzchaResponse
+extends EzchaPaginatedResponse
 class_name EzchaUserListResponse
-## A response from the API containing a single game.
+## A response from the API containing a paginated list of users.
 
 func _get_array_type_map() -> Dictionary:
 	return {
 		"users": EzchaUser
 	}
 
-## The users returned by the API request.
+## The list of users returned by the API request.
 var users: Array[EzchaUser] = []